% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FinalList}
\alias{FinalList}
\title{\code{InitialList} formatted for \code{shinyHMCI(...)}}
\format{A data frame with 9177 rows and 4 variables:
\describe{
  \item{x}{horizontal location}
  \item{y}{vertical location}
  \item{stat}{Statistic of interest}
  \item{lb}{Confidence level and location specific lower bound}
  \item{ub}{Confidence level and location specific lower bound}
}}
\usage{
FinalList
}
\description{
This list is \code{InitialList} formatted appropriately for \code{shinyHMCI(...)}.
}
\keyword{datasets}
