% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapper.R
\name{heatmapper}
\alias{heatmapper}
\title{Create a Simple Heat Map (by ggplot)}
\usage{
heatmapper(dataset, colors, plot_title = "Heat Map", guide_title = "Stat",
  lower = 0, upper = 0.17)
}
\arguments{
\item{dataset}{Data frame to use to build heat map. Coordinates for the heat map must be columns in \code{dataset} named \code{x} and \code{y}}

\item{colors}{The statistic of interest represented by a color at each (\code{x}, \code{y}) location on the heat map; must be saved in global environment, or supplied in this format: dataset$colors}

\item{plot_title}{The title for the plot, must be in quotes, defaults to "Heat Map"}

\item{guide_title}{The title for the guide, must be in quotes, defaults to "Stat"}

\item{lower}{lower bound for plot legend}

\item{upper}{upper bound for plot legend}
}
\value{
A heat map of the statistic supplied to the argument \code{colors}.
}
\description{
This function creates a basic heat map with ggplot(), for default use in shinyit().
}
\examples{
with(peralta[[1]], heatmapper(peralta[[1]], stat))
}
