% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hitter}
\alias{hitter}
\title{9177 swings by baseball player Jhonny Peralta}
\format{A data frame with 9177 rows and 4 variables:
\describe{
  \item{px}{horizontal location, in feet, from the middle of home plate when the pitch passed through the strike zone}
  \item{pz}{vertical location, in feet, from the ground when the pitch passed through the strike zone (negative values correspond to pitches that bounced before reaching home plate)}
  \item{des}{A short description of the swing outcome}
  \item{hit}{Bernoulli outcome: 1 if swing resulted in a hit, 0 otherwise}
}}
\source{
\url{http://www.baseballheatmaps.com/pitch-fx-download/}

\url{http://www.mlbam.com/}
}
\usage{
hitter
}
\description{
A dataset containing the location and outcome of 9177 Jhonny Peralta swings between 2008 and 2016.
}
\keyword{datasets}
