% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{InitialDF}
\alias{InitialDF}
\title{Data frame of spatial coordinates and variables}
\format{A data frame with 9177 rows and 4 variables:
\describe{
  \item{px}{horizontal location}
  \item{pz}{vertical location}
  \item{logit}{First model predicted value}
  \item{SE_logit}{Second model predicted value}
  \item{phat}{Third model predicted value}
}}
\usage{
InitialDF
}
\description{
A dataset containing spatial coordinates and three model predicted values for each location.
}
\keyword{datasets}
