% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{InitialList}
\alias{InitialList}
\title{List of 100 data frames}
\format{A data frame with 9177 rows and 4 variables:
\describe{
  \item{px}{horizontal location}
  \item{pz}{vertical location}
  \item{logit}{First model predicted value}
  \item{SE_logit}{Second model predicted value}
  \item{phat}{Third model predicted value}
  \item{lb}{Confidence level and location specific lower bound}
  \item{ub}{Confidence level and location specific lower bound}
}}
\usage{
InitialList
}
\description{
The first element is the \code{InitialDF} data frame. The remaining 99 data frames provide upper and lower bounds for each confidence level.
}
\keyword{datasets}
