% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapify.R
\name{mapify}
\alias{mapify}
\title{Create a Simple Heat Map (by ggplot)}
\usage{
mapify(dataset, stat)
}
\arguments{
\item{dataset}{Data frame to use to build heat map. Coordinates for the heat map must be columns in \code{dataset} named \code{x} and \code{y}}

\item{stat}{The statistic of interest represented by a color at each (\code{x}, \code{y}) location}
}
\value{
A heat map, representing \code{stat} with a color at each (\code{x}, \code{y}) location.
}
\description{
This function creates a basic heat map with ggplot(), for default use in shinyit().
}
\examples{
d0 <- data(CIdata0)
d1 <- as.data.frame(d0)
mapify(d1)
}
