% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peralta}
\alias{peralta}
\title{\code{peralta} formatted for \code{shinyHMCI(...)}}
\format{A lists of 100 data frames.  Element one is \code{\link{peralta_ests}}, elements 2-100 are \code{\link{peralta_cis}}.}
\usage{
peralta
}
\description{
This list combines \code{\link{peralta_ests}} and \code{\link{peralta_cis}} formatted appropriately for \code{shinyHMCI(...)}.
}
\keyword{datasets}
