% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peralta_cis}
\alias{peralta_cis}
\title{List of 99 data frames}
\format{A list of 99 data frames, the $i$th data frame corresponds to the $i$% CI with columns:
\describe{
  \item{lb}{Lower bound}
  \item{ub}{Upper bound}
}}
\usage{
peralta_cis
}
\description{
The 99 data frames provide upper and lower bounds for each confidence level associated with the point estimates in \code{\link{peralta_ests}}.
}
\keyword{datasets}
