% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyit.R
\name{shinyit}
\alias{shinyit}
\title{Create an Interactive Heat Map Confidence Interval (with Shiny)}
\usage{
shinyit(CIdata)
}
\arguments{
\item{CIdata}{A list with 100 data frames as elements. The first data frame element must contain: coordinates named \code{x} and \code{y}, and point estimate \code{stat}. The remaining 99 data frame elements consist of two columns each: lower bound, and upper bound.}
}
\value{
A heat map, representing \code{stat} with a color at each (\code{x}, \code{y}) location.
}
\description{
This function creates an interactive heat map confidence interval in Shiny.
}
\examples{
data(CIdata0)
mapify(CIdata0[[1]])
}
